+++ CHANGELOG.rst
@@ -36,6 +36,7 @@ Fixed
   must be called separately. (bug fix)
 * When inspecting custom YAQL/Jinja function to see if there is a context arg, use getargspec
   for py2 and getfullargspec for py3. (bug fix)
+* Check syntax on with items task to ensure action is indented correctly. Fixes #184 (bug fix)
 
 1.0.0
 -----